Move DuckBox to a renderer
